Compendiumization; go to 7fa01b3 for previous history
The analysis files and manual data files moved as a result of making this a reproducible compendium. Since GitHub doesn't use --follow in their logs, the history is hidden. Going to 7fa01b3 and seeing the history of /analysis_stuff.Rmd and Data/data_manual/* will show the whole past history